refactor: store all the revision related functions inside helper/revisions
